requirements.txt
fastapi
uvicorn
requests
pytz
gitpython 